+++ R/utils.R
@@ -44,7 +44,6 @@ label_aevar <- function(x) {
   return(out_label)
 }
 
-
 ci_choices <- setNames(c("wald", "waldcc", "ac", "scorecc", "score", "mn", "mee", "blj", "ha"),
                        c("Wald", "Corrected Wald", "Agresti-Caffo", "Newcombe",
                          "Score", "Miettinen and Nurminen", "Mee",
@@ -52,6 +51,7 @@ ci_choices <- setNames(c("wald", "waldcc", "ac", "scorecc", "score", "mn", "mee"
 
 #' retrieve name of ci method
 #' @param x ci method to retrieve its name
-name_ci <- function(x) {
-  names(ci_choices)[which(ci_choices == x)]
+name_ci <- function(x = ci_choices) {
+  x <- match.arg(x)
+  return(x)
 }